Fix bug on MsalService and finding a window for a login broker. => also created a local cache for msal login credentials so that if the token isn't expired it will automatically login in User.
The MainWindow login to microsoft is sort of different because the browser cache can store multiple users, and not necessary the user that creates the Cloud backups.
I think that some users are going to Use the browser for work, and don't want to login in to the FireBrowser tenant => so I made it work with different accounts on the MSALService, and browser auth.